pyproject.toml
[project]
name = "crm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-bootstrap5>=24.3",
    "django>=5.1.5",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "django-filter>=24.3",
    "django-widget-tweaks>=1.5.0",
]

[dependency-groups]
dev = [
    "ruff>=0.9.3",
]
